Extract shared font generation utilities into test/font-generation-helpers.mjs
The three generate-*.mjs test scripts duplicated binary helpers (u16, u32, i16,
etc.), checksum calculation, table builders (head, post), and font assembly
logic. This extracts the shared code into a single reusable module.

All generated font files remain byte-identical after this refactor.
