+++ test/generate-circular-ref-font.mjs
@@ -17,49 +17,11 @@
 import { writeFileSync } from 'fs';
 import { dirname, join } from 'path';
 import { fileURLToPath } from 'url';
+import { u16, u32, i16, i64, tag, pad, makeHead, makePost, assembleFont } from './font-generation-helpers.mjs';
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
 
-// --- helpers ---
-
-function u16(v) { return [(v >> 8) & 0xff, v & 0xff]; }
-function u32(v) { return [(v >> 24) & 0xff, (v >> 16) & 0xff, (v >> 8) & 0xff, v & 0xff]; }
-function i16(v) { return u16(v < 0 ? v + 0x10000 : v); }
-function i64(v) { return [...u32(0), ...u32(v)]; } // simplified LONGDATETIME
-function tag(s) { return [...s].map(c => c.charCodeAt(0)); }
-function pad(arr) { while (arr.length % 4 !== 0) arr.push(0); return arr; }
-
-function calcChecksum(bytes) {
-    const padded = [...bytes];
-    while (padded.length % 4 !== 0) padded.push(0);
-    let sum = 0;
-    for (let i = 0; i < padded.length; i += 4) {
-        sum = (sum + ((padded[i] << 24) | (padded[i+1] << 16) | (padded[i+2] << 8) | padded[i+3])) >>> 0;
-    }
-    return sum;
-}
-
 // --- table builders ---
-
-function makeHead() {
-    return [
-        ...u16(1), ...u16(0),          // majorVersion, minorVersion
-        ...u16(1), ...u16(0),          // fontRevision (fixed 1.0)
-        ...u32(0),                      // checksumAdjustment (filled later)
-        ...u32(0x5F0F3CF5),            // magicNumber
-        ...u16(0x000B),                // flags
-        ...u16(1000),                  // unitsPerEm
-        ...i64(0),                      // created
-        ...i64(0),                      // modified
-        ...i16(0), ...i16(0),          // xMin, yMin
-        ...i16(1000), ...i16(1000),    // xMax, yMax
-        ...u16(0),                      // macStyle
-        ...u16(8),                      // lowestRecPPEM
-        ...i16(2),                      // fontDirectionHint
-        ...i16(1),                      // indexToLocFormat (long)
-        ...i16(0),                      // glyphDataFormat
-    ];
-}
 
 function makeHhea(numGlyphs) {
     return [
@@ -175,20 +137,6 @@ function makeCmap() {
     ];
 }
 
-function makePost() {
-    return [
-        ...u16(3), ...u16(0),  // version 3.0 (no glyph names)
-        ...u32(0),              // italicAngle
-        ...i16(-100),          // underlinePosition
-        ...i16(50),            // underlineThickness
-        ...u32(0),              // isFixedPitch
-        ...u32(0),              // minMemType42
-        ...u32(0),              // maxMemType42
-        ...u32(0),              // minMemType1
-        ...u32(0),              // maxMemType1
-    ];
-}
-
 function makeName() {
     const names = [
         [0, 'Copyright'],
@@ -273,7 +221,7 @@ function buildFont() {
 
     const numGlyphs = 3;
     const tables = {
-        'head': makeHead(),
+        'head': makeHead({ indexToLocFormat: 1 }),
         'hhea': makeHhea(numGlyphs),
         'maxp': makeMaxp(numGlyphs),
         'OS/2': makeOs2(),
@@ -285,46 +233,7 @@ function buildFont() {
         'post': makePost(),
     };
 
-    const tags = Object.keys(tables).sort();
-    const numTables = tags.length;
-    const searchRange = Math.pow(2, Math.floor(Math.log2(numTables))) * 16;
-    const entrySelector = Math.floor(Math.log2(numTables));
-    const rangeShift = numTables * 16 - searchRange;
-
-    // Offset table (12 bytes) + table records (numTables * 16 bytes)
-    const headerSize = 12 + numTables * 16;
-    let dataOffset = headerSize;
-
-    // Build table records and collect padded table data
-    const tableRecords = [];
-    const tableData = [];
-    for (const t of tags) {
-        const data = tables[t];
-        const paddedData = pad([...data]);
-        tableRecords.push([
-            ...tag(t),
-            ...u32(calcChecksum(data)),
-            ...u32(dataOffset),
-            ...u32(data.length),
-        ]);
-        tableData.push(...paddedData);
-        dataOffset += paddedData.length;
-    }
-
-    const font = [
-        // Offset table
-        ...u32(0x00010000),         // sfVersion (TrueType)
-        ...u16(numTables),
-        ...u16(searchRange),
-        ...u16(entrySelector),
-        ...u16(rangeShift),
-        // Table records
-        ...tableRecords.flat(),
-        // Table data
-        ...tableData,
-    ];
-
-    return new Uint8Array(font);
+    return assembleFont(tables);
 }
 
 const fontBytes = buildFont();
