+++ test/generate-recursive-cff-font.mjs
@@ -10,18 +10,9 @@
  */
 
 import { writeFileSync } from 'fs';
+import { u8, u16, u32, i16, tag, makeHead, makePost, assembleFont } from './font-generation-helpers.mjs';
 
-// --- Byte-level helpers (big-endian) ---
-
-function u8(v) { return [v & 0xFF]; }
-function u16(v) { return [(v >> 8) & 0xFF, v & 0xFF]; }
-function u32(v) { return [(v >> 24) & 0xFF, (v >> 16) & 0xFF, (v >> 8) & 0xFF, v & 0xFF]; }
-function i16(v) { return u16(v < 0 ? v + 0x10000 : v); }
-function tag(s) { return [...s].map(c => c.charCodeAt(0)); }
-function pad4(bytes) {
-    while (bytes.length % 4 !== 0) bytes.push(0);
-    return bytes;
-}
+// --- CFF-specific helpers ---
 
 // CFF number encoding (Type 2 charstring format)
 function cffInt(v) {
@@ -148,28 +139,6 @@ function buildCFF() {
 
 function buildOTF() {
     const numGlyphs = 2;
-    const unitsPerEm = 1000;
-
-    // --- Required tables ---
-
-    // head table (54 bytes)
-    const headTable = [
-        ...u16(1), ...u16(0),          // version 1.0
-        ...u16(1), ...u16(0),          // fontRevision 1.0
-        ...u32(0),                      // checksumAdjustment (filled later)
-        ...u32(0x5F0F3CF5),            // magicNumber
-        ...u16(0x000B),                // flags
-        ...u16(unitsPerEm),            // unitsPerEm
-        ...u32(0), ...u32(0),          // created (longDateTime)
-        ...u32(0), ...u32(0),          // modified (longDateTime)
-        ...i16(0), ...i16(0),          // xMin, yMin
-        ...i16(1000), ...i16(1000),    // xMax, yMax
-        ...u16(0),                      // macStyle
-        ...u16(8),                      // lowestRecPPEM
-        ...i16(2),                      // fontDirectionHint
-        ...i16(1),                      // indexToLocFormat
-        ...i16(0),                      // glyphDataFormat
-    ];
 
     // hhea table (36 bytes)
     const hheaTable = [
@@ -220,90 +189,11 @@ function buildOTF() {
         ...u32(0),                      // ulCodePageRange2
     ];
 
-    // name table (minimal - just required nameIDs with short strings)
-    function buildNameTable() {
-        const names = [
-            [0, 'Copyright'],
-            [1, 'Test'],
-            [2, 'Regular'],
-            [4, 'Test'],
-            [5, 'Version 1.0'],
-            [6, 'Test-Regular'],
-        ];
-        const stringData = [];
-        const records = [];
-        let offset = 0;
-        for (const [nameID, str] of names) {
-            // Platform 3 (Windows), encoding 1 (Unicode BMP), language 0x0409 (English)
-            const encoded = [];
-            for (const ch of str) {
-                encoded.push(0, ch.charCodeAt(0));
-            }
-            records.push([3, 1, 0x0409, nameID, encoded.length, offset]);
-            stringData.push(...encoded);
-            offset += encoded.length;
-        }
-        const count = records.length;
-        const storageOffset = 6 + count * 12;
-        const result = [...u16(0), ...u16(count), ...u16(storageOffset)];
-        for (const [platID, encID, langID, nameID, len, off] of records) {
-            result.push(...u16(platID), ...u16(encID), ...u16(langID), ...u16(nameID), ...u16(len), ...u16(off));
-        }
-        result.push(...stringData);
-        return result;
-    }
+    // name table
     const nameTable = buildNameTable();
 
     // cmap table (format 4, maps space U+0020 to glyph 1)
-    function buildCmapTable() {
-        const segCount = 2; // one real segment + sentinel
-        const searchRange = 2 * Math.pow(2, Math.floor(Math.log2(segCount)));
-        const entrySelector = Math.floor(Math.log2(segCount));
-        const rangeShift = 2 * segCount - searchRange;
-
-        const subtable = [
-            ...u16(4),                     // format
-            ...u16(0),                     // length (filled below)
-            ...u16(0),                     // language
-            ...u16(segCount * 2),          // segCountX2
-            ...u16(searchRange),
-            ...u16(entrySelector),
-            ...u16(rangeShift),
-            // endCode
-            ...u16(0x0020), ...u16(0xFFFF),
-            // reservedPad
-            ...u16(0),
-            // startCode
-            ...u16(0x0020), ...u16(0xFFFF),
-            // idDelta
-            ...i16(1 - 0x0020), ...i16(1),
-            // idRangeOffset
-            ...u16(0), ...u16(0),
-        ];
-        // Fix length
-        subtable[2] = (subtable.length >> 8) & 0xFF;
-        subtable[3] = subtable.length & 0xFF;
-
-        return [
-            ...u16(0),                     // version
-            ...u16(1),                     // numTables
-            ...u16(3), ...u16(1),          // platformID=3, encodingID=1
-            ...u32(12),                    // offset to subtable
-            ...subtable,
-        ];
-    }
     const cmapTable = buildCmapTable();
-
-    // post table (format 3 - no glyph names, 32 bytes)
-    const postTable = [
-        ...u16(0x0003), ...u16(0x0000), // format 3.0
-        ...u16(0), ...u16(0),           // italicAngle
-        ...i16(-100),                    // underlinePosition
-        ...i16(50),                      // underlineThickness
-        ...u32(0),                       // isFixedPitch
-        ...u32(0), ...u32(0),           // minMemType42, maxMemType42
-        ...u32(0), ...u32(0),           // minMemType1, maxMemType1
-    ];
 
     // hmtx table
     const hmtxTable = [
@@ -314,70 +204,89 @@ function buildOTF() {
     // CFF table
     const cffTable = buildCFF();
 
-    // --- Assemble OTF ---
-    const tables = [
-        ['CFF ', cffTable],
-        ['OS/2', os2Table],
-        ['cmap', cmapTable],
-        ['head', headTable],
-        ['hhea', hheaTable],
-        ['hmtx', hmtxTable],
-        ['maxp', maxpTable],
-        ['name', nameTable],
-        ['post', postTable],
+    const tables = {
+        'CFF ': cffTable,
+        'OS/2': os2Table,
+        'cmap': cmapTable,
+        'head': makeHead({ indexToLocFormat: 1 }),
+        'hhea': hheaTable,
+        'hmtx': hmtxTable,
+        'maxp': maxpTable,
+        'name': nameTable,
+        'post': makePost(),
+    };
+
+    return assembleFont(tables, { sfVersion: 'OTTO' });
+}
+
+function buildNameTable() {
+    const names = [
+        [0, 'Copyright'],
+        [1, 'Test'],
+        [2, 'Regular'],
+        [4, 'Test'],
+        [5, 'Version 1.0'],
+        [6, 'Test-Regular'],
     ];
-
-    const numTables = tables.length;
-    const searchRange2 = Math.pow(2, Math.floor(Math.log2(numTables))) * 16;
-    const entrySelector2 = Math.floor(Math.log2(numTables));
-    const rangeShift2 = numTables * 16 - searchRange2;
-
-    // OTF header
-    const sfntHeader = [
-        ...tag('OTTO'),                 // sfVersion (CFF)
-        ...u16(numTables),
-        ...u16(searchRange2),
-        ...u16(entrySelector2),
-        ...u16(rangeShift2),
-    ];
-
-    // Calculate table offsets
-    const tableRecordSize = 16; // tag(4) + checksum(4) + offset(4) + length(4)
-    let dataOffset = sfntHeader.length + numTables * tableRecordSize;
-
-    const tableEntries = [];
-    for (const [name, data] of tables) {
-        const paddedLen = Math.ceil(data.length / 4) * 4;
-        tableEntries.push({ tag: name, data, offset: dataOffset, length: data.length });
-        dataOffset += paddedLen;
-    }
-
-    // Build table records
-    function checksum(bytes) {
-        let sum = 0;
-        const padded = [...bytes];
-        while (padded.length % 4 !== 0) padded.push(0);
-        for (let i = 0; i < padded.length; i += 4) {
-            sum = (sum + ((padded[i] << 24) | (padded[i+1] << 16) | (padded[i+2] << 8) | padded[i+3])) >>> 0;
-        }
-        return sum;
-    }
-
+    const stringData = [];
     const records = [];
-    for (const entry of tableEntries) {
-        records.push(...tag(entry.tag));
-        records.push(...u32(checksum(entry.data)));
-        records.push(...u32(entry.offset));
-        records.push(...u32(entry.length));
+    let offset = 0;
+    for (const [nameID, str] of names) {
+        // Platform 3 (Windows), encoding 1 (Unicode BMP), language 0x0409 (English)
+        const encoded = [];
+        for (const ch of str) {
+            encoded.push(0, ch.charCodeAt(0));
+        }
+        records.push([3, 1, 0x0409, nameID, encoded.length, offset]);
+        stringData.push(...encoded);
+        offset += encoded.length;
     }
-
-    // Assemble final file
-    const file = [...sfntHeader, ...records];
-    for (const entry of tableEntries) {
-        file.push(...pad4([...entry.data]));
+    const count = records.length;
+    const storageOffset = 6 + count * 12;
+    const result = [...u16(0), ...u16(count), ...u16(storageOffset)];
+    for (const [platID, encID, langID, nameID, len, off] of records) {
+        result.push(...u16(platID), ...u16(encID), ...u16(langID), ...u16(nameID), ...u16(len), ...u16(off));
     }
+    result.push(...stringData);
+    return result;
+}
 
-    return new Uint8Array(file);
+function buildCmapTable() {
+    const segCount = 2; // one real segment + sentinel
+    const searchRange = 2 * Math.pow(2, Math.floor(Math.log2(segCount)));
+    const entrySelector = Math.floor(Math.log2(segCount));
+    const rangeShift = 2 * segCount - searchRange;
+
+    const subtable = [
+        ...u16(4),                     // format
+        ...u16(0),                     // length (filled below)
+        ...u16(0),                     // language
+        ...u16(segCount * 2),          // segCountX2
+        ...u16(searchRange),
+        ...u16(entrySelector),
+        ...u16(rangeShift),
+        // endCode
+        ...u16(0x0020), ...u16(0xFFFF),
+        // reservedPad
+        ...u16(0),
+        // startCode
+        ...u16(0x0020), ...u16(0xFFFF),
+        // idDelta
+        ...i16(1 - 0x0020), ...i16(1),
+        // idRangeOffset
+        ...u16(0), ...u16(0),
+    ];
+    // Fix length
+    subtable[2] = (subtable.length >> 8) & 0xFF;
+    subtable[3] = subtable.length & 0xFF;
+
+    return [
+        ...u16(0),                     // version
+        ...u16(1),                     // numTables
+        ...u16(3), ...u16(1),          // platformID=3, encodingID=1
+        ...u32(12),                    // offset to subtable
+        ...subtable,
+    ];
 }
 
 const otf = buildOTF();
